+++ deploy/utils/addresses_metis.js
@@ -1,0 +1,7 @@
+module.exports = {
+  // Deployed addresses
+
+  // External addresses
+  HUMMUS_ROUTER01: '0x6B6F7437DF9cE9552ED7Fc8f529BAf48fb305534',
+};
+module.exports.skip = async () => true;